+++ headers/minishell.h
@@ -6,17 +6,21 @@
 /*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 10:53:20 by umeneses          #+#    #+#             */
-/*   Updated: 2024/06/12 10:53:23 by umeneses         ###   ########.fr       */
+/*   Updated: 2024/06/12 14:15:14 by umeneses         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
+# include "../libs/libft/includes/libft.h"
 # include <stdio.h>
 # include <stdlib.h>
 # include <readline/readline.h>
 # include <readline/history.h>
+# include <sys/wait.h>
+# include <errno.h>
+# include <unistd.h>
 
 # define RED     "\x1b[31m"
 # define GREEN   "\x1b[32m"
