+++ libs/libft/src/ft_putchar_fd.c
@@ -1,34 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putnbr_unsg_fd.c                                :+:      :+:    :+:   */
+/*   ft_putchar_fd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmalheir <tmalheir@student.42.fr>          +#+  +:+       +#+        */
+/*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/06 09:11:02 by tmalheir          #+#    #+#             */
-/*   Updated: 2024/06/11 15:02:26 by tmalheir         ###   ########.fr       */
+/*   Updated: 2024/06/12 14:15:56 by umeneses         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/ft_printf.h"
+#include "../includes/libft.h"
 
-unsigned int	ft_putnbr_unsg_fd(unsigned int nbr, int fd)
+int	ft_putchar_fd(char c, int fd)
 {
-	char	n_char;
-	long	n_long;
-	int		count;
+	size_t	count;
 
-	n_long = nbr;
-	count = 0;
-	if (n_long >= 0 && n_long <= 9)
-	{
-		n_char = n_long + '0';
-		count += write(1, &n_char, 1);
-	}
-	else
-	{
-		count += ft_putnbr_unsg_fd((n_long / 10), fd);
-		count += ft_putnbr_unsg_fd((n_long % 10), fd);
-	}
+	count = write(1, &c, fd);
 	return (count);
 }
