+++ libs/libft/src/ft_puthexa_low_fd.c
@@ -1,21 +1,30 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putchar_fd.c                                    :+:      :+:    :+:   */
+/*   ft_puthexa_low_fd.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmalheir <tmalheir@student.42.fr>          +#+  +:+       +#+        */
+/*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/06 09:11:02 by tmalheir          #+#    #+#             */
-/*   Updated: 2024/06/11 15:02:12 by tmalheir         ###   ########.fr       */
+/*   Updated: 2024/06/12 14:15:59 by umeneses         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/ft_printf.h"
+#include "../includes/libft.h"
 
-int	ft_putchar_fd(char c, int fd)
+int	ft_puthexa_low_fd(unsigned long long nbr, int fd)
 {
-	size_t	count;
+	int		count;
+	char	*hexa_low;
 
-	count = write(1, &c, fd);
+	count = 0;
+	hexa_low = "0123456789abcdef";
+	if (nbr < 16)
+		count += write(1, &hexa_low[nbr], 1);
+	else
+	{
+		count += ft_puthexa_low_fd((nbr / 16), fd);
+		count += ft_puthexa_low_fd((nbr % 16), fd);
+	}
 	return (count);
 }
