+++ libs/libft/src/ft_putnbr_unsg_fd.c
@@ -1,31 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putstr_fd.c                                     :+:      :+:    :+:   */
+/*   ft_putnbr_unsg_fd.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmalheir <tmalheir@student.42.fr>          +#+  +:+       +#+        */
+/*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/06 09:11:02 by tmalheir          #+#    #+#             */
-/*   Updated: 2024/06/11 15:02:32 by tmalheir         ###   ########.fr       */
+/*   Updated: 2024/06/12 14:16:07 by umeneses         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/ft_printf.h"
+#include "../includes/libft.h"
 
-int	ft_putstr_fd(char *str, int fd)
+unsigned int	ft_putnbr_unsg_fd(unsigned int nbr, int fd)
 {
-	int	idx;
+	char	n_char;
+	long	n_long;
+	int		count;
 
-	idx = 0;
-	if (!str)
-		return (write(1, "(null)", 6));
+	n_long = nbr;
+	count = 0;
+	if (n_long >= 0 && n_long <= 9)
+	{
+		n_char = n_long + '0';
+		count += write(1, &n_char, 1);
+	}
 	else
 	{
-		while (str && str[idx])
-		{
-			write(1, &str[idx], fd);
-			idx++;
-		}
+		count += ft_putnbr_unsg_fd((n_long / 10), fd);
+		count += ft_putnbr_unsg_fd((n_long % 10), fd);
 	}
-	return (idx);
+	return (count);
 }
