+++ libs/libft/src/ft_putptr_fd.c
@@ -1,39 +1,36 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
+/*   ft_putptr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmalheir <tmalheir@student.42.fr>          +#+  +:+       +#+        */
+/*   By: umeneses <umeneses@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/06 09:11:02 by tmalheir          #+#    #+#             */
-/*   Updated: 2024/06/11 15:02:23 by tmalheir         ###   ########.fr       */
+/*   Updated: 2024/06/12 14:16:09 by umeneses         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/ft_printf.h"
+#include "../includes/libft.h"
 
-int	ft_putnbr_fd(int nbr, int fd)
+int	ft_putptr_fd(unsigned long long ptr, int fd)
 {
-	char	n_char;
-	long	n_long;
 	int		count;
+	char	*hexa_low;
 
-	n_long = nbr;
 	count = 0;
-	if (n_long < 0)
+	hexa_low = "0123456789abcdef";
+	if (!ptr)
 	{
-		count += write(1, "-", 1);
-		n_long = -n_long;
+		count += write(1, "(nil)", 5);
+		return (count);
 	}
-	if (n_long >= 0 && n_long <= 9)
-	{
-		n_char = n_long + '0';
-		count += write(1, &n_char, 1);
-	}
+	count += write(1, "0x", 2);
+	if (ptr < 16)
+		count += write(1, &hexa_low[ptr], 1);
 	else
 	{
-		count += ft_putnbr_fd((n_long / 10), fd);
-		count += ft_putnbr_fd((n_long % 10), fd);
+		count += ft_puthexa_low_fd((ptr / 16), fd);
+		count += ft_puthexa_low_fd((ptr % 16), fd);
 	}
 	return (count);
 }
